requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy==1.23.5
opencv-python>=4.10.0.84
opencv-contrib-python>=4.10.0.84
opencv-python-headless>=4.10.0.84
transformers>=4.30.0,<4.41.0
diffusers>=0.25.0
accelerate>=0.25.0
bitsandbytes>=0.39.0
deepspeed>=0.14.2
huggingface-hub>=0.24.5
lightning>=2.2.4
lightning-utilities>=0.11.6
controlnet-aux>=0.0.9
ffmpeg-python
einops
imageio
moviepy
decord
av
librosa
pydub
pandas
scikit-learn
scikit-image
matplotlib
seaborn
albumentations
basicsr
facexlib
insightface
gfpgan>=1.3.8
PyYAML
python-dotenv
hydra-core
omegaconf
requests
pillow
opentelemetry-api
opentelemetry-sdk
black
ipython
ipdb
jupyterlab_widgets
fastapi
fastapi-cli
flask
flask-cors
gradio
gradio_client
gunicorn
lpips
trimesh
plyfile
open3d
smplx
tqdm
loguru
natsort
easydict